Fixed some bugs 07/21/2022 | 22w29b1

+++ TRAP/src/Network/FTP/FTP.cpp
@@ -559,7 +559,7 @@ TRAP::Network::FTP::Response TRAP::Network::FTP::DataChannel::Open(const Transfe
 			}
 
 			//Reconstruct connection port and address
-			const uint16_t port = static_cast<uint8_t>(data[4] * 256) + data[5];
+			const uint16_t port = static_cast<uint16_t>(data[4] * 256) + data[5];
 			const IPv4Address address(data[0], data[1], data[2], data[3]);
 
 			//Connect the data channel to the server
